requirements.txt
ultralytics>=8.0.0
opencv-python>=4.8.0
numpy>=1.23.0
pandas>=2.0.0
scikit-learn>=1.3.0
fastapi>=0.110.0
uvicorn>=0.29.0
pydantic>=2.6.0
pyyaml>=6.0.1
streamlit>=1.33.0
tqdm>=4.66.0
